handle serial buffer overflow condition (while the modem protocol *can* send frames >512 bytes the serial port handler can only handle frames up to 512 bytes); better handle transmitting double length (512 byte) P25 PDU frames;

+++ p25/P25TX.cpp
@@ -151,7 +151,7 @@ void P25TX::process()
 
 /* Write data to the local buffer. */
 
-uint8_t P25TX::writeData(const uint8_t* data, uint8_t length)
+uint8_t P25TX::writeData(const uint8_t* data, uint16_t length)
 {
     if (length < (P25_TDU_FRAME_LENGTH_BYTES + 1U))
         return RSN_ILLEGAL_LENGTH;
@@ -164,7 +164,7 @@ uint8_t P25TX::writeData(const uint8_t* data, uint8_t length)
     }
 
     m_fifo.put(length - 1U);
-    for (uint8_t i = 0U; i < (length - 1U); i++)
+    for (uint16_t i = 0U; i < (length - 1U); i++)
         m_fifo.put(data[i + 1U]);
 
     return RSN_OK;
@@ -257,9 +257,9 @@ void P25TX::createData()
             m_poBuffer[m_poLen++] = P25_START_SYNC;
     }
     else {
-        uint8_t length = m_fifo.get();
+        uint16_t length = m_fifo.get();
         DEBUG3("P25TX::createData() dataLength/fifoSpace", length, m_fifo.getSpace());
-        for (uint8_t i = 0U; i < length; i++) {
+        for (uint16_t i = 0U; i < length; i++) {
             m_poBuffer[m_poLen++] = m_fifo.get();
         }
     }
@@ -273,7 +273,7 @@ void P25TX::createCal()
 {
     // 1.2 kHz sine wave generation
     if (m_modemState == STATE_P25_CAL) {
-        for (unsigned int i = 0U; i < P25_LDU_FRAME_LENGTH_BYTES; i++) {
+        for (uint8_t i = 0U; i < P25_LDU_FRAME_LENGTH_BYTES; i++) {
             m_poBuffer[i] = P25_START_SYNC;
         }
